tests/sensors/AbsoluteOrientationSensor.https.html

<!DOCTYPE html>
<!--
Copyright (c) 2017 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<title>AbsoluteOrientationSensor Test</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Xiuqi Jiang" href="mailto:xiuqix.jiang@intel.com">
<link rel="help" href="https://w3c.github.io/orientation-sensor/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./resources/generic-sensor-tests.js"></script>
<div id="log"></div>
<iframe src="./resources/AbsoluteOrientationSensor-support-iframe.html" id="frame" style="display:none" sandbox="allow-scripts">
</iframe>
<script>

const quat_init = [0, 1, 0, 0];
const mat_init = [1,  0,  0,  0,
                  0, -1,  0,  0,
                  0,  0, -1,  0,
                  0,  0,  0,  1];

function create_matrix(quat) {
  let W = quat[0];
  let X = quat[1];
  let Y = quat[2];
  let Z = quat[3];
  let mat = new Array(
    1-2*Y*Y-2*Z*Z, 2*X*Y-2*Z*W, 2*X*Z+2*W*Y, 0,
    2*X*Y+2*W*Z, 1-2*X*X-2*Z*Z, 2*Y*Z-2*W*X, 0,
    2*X*Z-2*W*Y, 2*Y*Z+2*W*X, 1-2*X*X-2*Y*Y, 0,
    0, 0, 0, 1
  );
  return mat;
}

async_test(t => {
  let sensor = new AbsoluteOrientationSensor();
  sensor.onchange = t.step_func_done(() => {
    assert_equals(sensor.quaternion.length, 4);
    assert_true(sensor.quaternion instanceof Array);
    sensor.stop();
  });
  sensor.onerror = t.step_func_done(unreached);
  sensor.start();
}, "sensor.quaternion return a four-element FrozenArray");

async_test(t => {
  let sensor = new AbsoluteOrientationSensor();
  sensor.onchange = t.step_func_done(() => {
    assert_throws({ name: 'SyntaxError' }, () => sensor.populateMatrix("not RotationMatrixType parameter"), "not RotationMatrixType parameter");
    assert_throws({ name: 'TypeError' }, () => sensor.populateMatrix(new Float32Array(15)), "a size less than sixteen Float32Array type parameter");
    sensor.stop();
  });
  sensor.onerror = t.step_func_done(unreached);
  sensor.start();
}, "exception is thrown when 'populateMatrix' event has a wrong paramter");

async_test(t => {
  let sensor = new AbsoluteOrientationSensor();
  let mat_32 = new Float32Array(16);
  sensor.onchange = t.step_func_done(() => {
    let quat = sensor.quaternion;
    sensor.populateMatrix(mat_32);
    let mat_com = create_matrix(quat);
    assert_array_equals(mat_32, mat_com);
    sensor.stop();
  });
  sensor.onerror = t.step_func_done(unreached);
  sensor.start();
}, "the data produced from Float32Array populateMatrix function is correct");

async_test(t => {
  let sensor = new AbsoluteOrientationSensor();
  let mat_64 = new Float64Array(16);
  sensor.onchange = t.step_func_done(() => {
    let quat = sensor.quaternion;
    sensor.populateMatrix(mat_64);
    let mat_com = create_matrix(quat);
    assert_array_equals(mat_64, mat_com);
    sensor.stop();
  });
  sensor.onerror = t.step_func_done(unreached);
  sensor.start();
}, "the data produced from Float64Array populateMatrix function is correct");

async_test(t => {
  let sensor = new AbsoluteOrientationSensor();
  let mat_dom = new DOMMatrix(mat_init);
  sensor.onchange = t.step_func_done(() => {
    let quat = sensor.quaternion;
    sensor.populateMatrix(mat_dom);
    let mat_com = create_matrix(quat);
    assert_equals(mat_dom.m11, mat_com[0]);
    assert_equals(mat_dom.m12, mat_com[1]);
    assert_equals(mat_dom.m13, mat_com[2]);
    assert_equals(mat_dom.m14, mat_com[3]);
    assert_equals(mat_dom.m21, mat_com[4]);
    assert_equals(mat_dom.m22, mat_com[5]);
    assert_equals(mat_dom.m23, mat_com[6]);
    assert_equals(mat_dom.m24, mat_com[7]);
    assert_equals(mat_dom.m31, mat_com[8]);
    assert_equals(mat_dom.m32, mat_com[9]);
    assert_equals(mat_dom.m33, mat_com[10]);
    assert_equals(mat_dom.m34, mat_com[11]);
    assert_equals(mat_dom.m41, mat_com[12]);
    assert_equals(mat_dom.m42, mat_com[13]);
    assert_equals(mat_dom.m43, mat_com[14]);
    assert_equals(mat_dom.m44, mat_com[15]);
    sensor.stop();
  });
  sensor.onerror = t.step_func_done(unreached);
  sensor.start();
}, "the data produced from DOMMatrix populateMatrix function is correct");

function verify_sensor_reading(sensor, reading_exist = true) {
  if (reading_exist) {
    if (sensor.timestamp == null || sensor.quaternion.toString() == quat_init.toString()) {
      return false;
    }
    return true;
  }
  return sensor.quaternion == null && sensor.timestamp == null;
}

function reading_to_array(sensor) {
  let mat = new Float32Array(16);
  sensor.populateMatrix(mat);
  return mat;
}

runSensorFrequency(AbsoluteOrientationSensor);
runGenericSensorTests(AbsoluteOrientationSensor, verify_sensor_reading, reading_to_array);

</script>
